refactor(context menu): move context menu controls out to own hook
