+++ src/components/Files.tsx
@@ -3,6 +3,7 @@ import { useOutletContext } from 'react-router';
 
 import { File } from '../shared.types';
 import useDisplayOptions from '../hooks/useDisplayOptions';
+import useContextMenu from '../hooks/useContextMenu';
 import FileList from './ui/FileList';
 import FilePropertiesDrawer from './ui/FilePropertiesDrawer';
 import FileControlPanel from './ui/FileControlPanel';
@@ -21,18 +22,24 @@ export default function Files() {
 
   const {
     propertiesTarget,
+    setPropertiesTarget,
     selectedFiles,
+    setSelectedFiles,
     displayFiles,
     hideDotFiles,
     setHideDotFiles,
     showFileDrawer,
     setShowFileDrawer,
-    showFileContextMenu,
-    setShowFileContextMenu,
-    contextMenuCoords,
-    handleContextMenu,
     handleLeftClicks
   } = useDisplayOptions(files);
+
+  const {
+    contextMenuCoords,
+    showFileContextMenu,
+    setShowFileContextMenu,
+    menuRef,
+    handleRightClick
+  } = useContextMenu();
 
   return (
     <div className="flex-1 overflow-auto flex flex-col">
@@ -50,18 +57,20 @@ export default function Files() {
           displayFiles={displayFiles}
           currentPath={currentPath}
           selectedFiles={selectedFiles}
+          setSelectedFiles={setSelectedFiles}
           selectedZone={selectedZone}
           getFiles={getFiles}
-          handleContextMenu={handleContextMenu}
           handleLeftClicks={handleLeftClicks}
           showFileDrawer={showFileDrawer}
+          setPropertiesTarget={setPropertiesTarget}
+          handleRightClick={handleRightClick}
         />
       </div>
       {showFileContextMenu && (
         <FileContextMenu
           x={contextMenuCoords.x}
           y={contextMenuCoords.y}
-          onClose={() => setShowFileContextMenu(false)}
+          menuRef={menuRef}
           setShowFileDrawer={setShowFileDrawer}
           setShowFileContextMenu={setShowFileContextMenu}
         />
