+++ src/components/ui/FileContextMenu.tsx
@@ -5,7 +5,7 @@ import { Typography } from '@material-tailwind/react';
 type ContextMenuProps = {
   x: number;
   y: number;
-  onClose: () => void;
+  menuRef: React.RefObject<HTMLDivElement>;
   setShowFileDrawer: (show: boolean) => void;
   setShowFileContextMenu: (show: boolean) => void;
 };
@@ -13,47 +13,10 @@ type ContextMenuProps = {
 export default function FileContextMenu({
   x,
   y,
-  onClose,
+  menuRef,
   setShowFileDrawer,
   setShowFileContextMenu
 }: ContextMenuProps): JSX.Element {
-  const menuRef = React.useRef<HTMLDivElement>(null);
-
-  React.useEffect(() => {
-    // Adjust menu position if it would go off screen
-    if (menuRef.current) {
-      const rect = menuRef.current.getBoundingClientRect();
-      const viewportWidth = window.innerWidth;
-      const viewportHeight = window.innerHeight;
-
-      let adjustedX = x;
-      let adjustedY = y;
-
-      if (x + rect.width > viewportWidth) {
-        adjustedX = viewportWidth - rect.width - 5;
-      }
-
-      if (y + rect.height > viewportHeight) {
-        adjustedY = viewportHeight - rect.height - 5;
-      }
-
-      menuRef.current.style.left = `${adjustedX}px`;
-      menuRef.current.style.top = `${adjustedY}px`;
-    }
-
-    // Add click handler to close the menu when clicking outside
-    const handleClickOutside = (e: MouseEvent) => {
-      if (menuRef.current && !menuRef.current.contains(e.target as Node)) {
-        onClose();
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [x, y, onClose]);
-
   return ReactDOM.createPortal(
     <div
       ref={menuRef}
